backend/chatbot/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
cohere==5.11.3
pydantic==2.9.2
pydantic-settings==2.6.1
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
httpx==0.27.2